+++ lib/canopus/project/search_worker/cancelled.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-class Canopus::Project::SearchWorker::Cancelled < StandardError; end